refactor: create `content/definitions/Post`
Remove Post definitions from `contentlayer.config.js`
