+++ contentlayer.config.js
@@ -1,49 +1,16 @@
+import remarkGfm from 'remark-gfm'
 import rehypePrettyCode from 'rehype-pretty-code'
 
-import { defineDocumentType, makeSource } from 'contentlayer/source-files'
-import { rehypePrettyCodeOptions } from './src/lib/rehypePrettyCode'
+import { makeSource } from 'contentlayer/source-files'
 
-export const Post = defineDocumentType(() => ({
-  name: 'Post',
-  contentType: 'mdx',
-  filePathPattern: 'posts/*.mdx',
-  fields: {
-    title: {
-      type: 'string',
-      required: true
-    },
-    date: {
-      type: 'date',
-      required: true
-    },
-    lastUpdate: {
-      type: 'date'
-    },
-    description: {
-      type: 'string',
-      required: true
-    },
-    category: {
-      type: 'string',
-      required: true
-    },
-    tags: {
-      type: 'string',
-      required: true
-    }
-  },
-  computedFields: {
-    id: {
-      type: 'string',
-      resolve: post => post._raw.sourceFileName.replace(/\.mdx$/, '')
-    }
-  }
-}))
+import { Post } from './content/definitions/Post'
+import { rehypePrettyCodeOptions } from './src/lib/rehypePrettyCode'
 
 export default makeSource({
   contentDirPath: 'content',
   documentTypes: [Post],
   mdx: {
+    remarkPlugins: [[remarkGfm]],
     rehypePlugins: [[rehypePrettyCode, rehypePrettyCodeOptions]]
   }
 })
